respuestas_estructuradas/requirements.txt
# Respuestas estructuradas: Pydantic AI + Google Gemini
pydantic-ai[google]>=0.0.14
google-genai>=1.0.0
python-dotenv>=1.0.0
